Setting the reduce reads count tag was all wrong in a previous commit; fixing.
RR counts are represented as offsets from the first count, but that wasn't being done
correctly when counts are adjusted on the fly.  Also, we were triggering the expensive
conversion and writing to binary tags even when we weren't going to write the read
to disk.

The code has been updated so that unconverted counts are passed to the GATKSAMRecord
and it knows how to encode the tag correctly.  Also, there are now methods to write
to the reduced counts array without forcing the conversion (and methods that do force
the conversion).

Also:
1. counts are now maintained as ints whenever possible.  Only the GATKSAMRecord knows
about the internal encoding.
2. as discussed in meetings today, we updated the encoding so that it can now handle
a range of values that extends to 255 instead of 127 (and is backwards compatible).
3. tests have been moved from SyntheticReadUnitTest to GATKSAMRecordUnitTest accordingly.

+++ public/java/src/org/broadinstitute/sting/utils/clipping/ClippingOp.java
@@ -380,7 +380,7 @@ public class ClippingOp {
         }
 
         if (read.isReducedRead()) {
-            final byte[] reducedCounts = new byte[newLength];
+            final int[] reducedCounts = new int[newLength];
             System.arraycopy(read.getReducedReadCounts(), copyStart, reducedCounts, 0, newLength);
             hardClippedRead.setReducedReadCounts(reducedCounts);
         }
